Move explorer includes into a folder
This will prevent confusion with details/content hub prototype includes
