+++ explorer/index.php
@@ -3,9 +3,9 @@
 <main>
     <?php require_once '../includes/header-and-nav-bar.php' ?>
     <div class="container">
-        <?php require_once('../includes/home-page-cards.php'); ?>
-        <?php require_once('../includes/popular-today.php'); ?>
-        <?php require_once('../includes/featured.php'); ?>
+        <?php require_once('../includes/explorer/home-page-cards.php'); ?>
+        <?php require_once('../includes/explorer/popular-today.php'); ?>
+        <?php require_once('../includes/explorer/featured.php'); ?>
     </div>
 </main>
 <?php require_once '../includes/footer.php' ?>
